Add moody_html module to avoid conflicts from flex_html upstream changes
Our usage of custom styles in ckeditor for Moody sites has been presenting an issue when flex_html changes are introduced in the upstream. We’ll maintain an eye on flex_html upstream changed to absorb anything needed for optimal site editing, and when needed we can migrate existing content over to moody_html when upstream flex_html conflicts arise.